requirements.txt
accelerate==1.1.1
aiofiles==23.2.1
aiohappyeyeballs==2.4.3
aiohttp==3.11.2
aiosignal==1.3.1
alabaster==0.7.16
annotated-types==0.7.0
anyio==4.6.2.post1
appdirs==1.4.4
APScheduler==3.10.4
asttokens==2.4.1
async-timeout==4.0.3
attrs==24.2.0
auditnlg==0.0.1
babel==2.16.0
beautifulsoup4==4.12.3
bitsandbytes==0.42.0
black==24.10.0
blis==0.7.11
blobfile==3.0.0
boltons==21.0.0
boto3==1.35.62
botocore==1.35.62
bracex==2.5.post1
Brotli==1.1.0
cachetools==5.5.0
catalogue==2.0.10
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
click-option-group==0.5.6
cloudpathlib==0.20.0
codeshield==1.0.1
colorama==0.4.6
coloredlogs==15.0.1
commonmark==0.9.1
confection==0.1.5
contourpy==1.3.1
cycler==0.12.1
cymem==2.0.8
datasets==3.1.0
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
detoxify==0.5.2
dill==0.3.8
distro==1.9.0
docformatter==1.7.5
docutils==0.15.2
edlib==1.3.9.post1
emoji==2.14.0
exceptiongroup==1.2.2
executing==2.1.0
face==24.0.0
fastapi==0.115.5
ffmpy==0.4.0
filelock==3.16.1
fire==0.7.0
flake8==7.1.1
flake8-bugbear==24.10.31
fonttools==4.55.0
frozenlist==1.5.0
fsspec==2024.9.0
gdown==5.2.0
gitdb==4.0.11
gitdb2==4.0.2
GitPython==3.1.43
glom==22.1.0
google-api-core==2.23.0
google-api-python-client==2.153.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.66.0
gradio==5.5.0
gradio_client==1.4.2
groq==0.4.0
h11==0.14.0
httpcore==1.0.6
httplib2==0.22.0
httpx==0.27.2
huggingface==0.0.1
huggingface-hub==0.26.2
humanfriendly==10.0
idna==3.10
imagesize==1.4.1
importlib_metadata==7.1.0
inflate64==1.0.0
iniconfig==2.0.0
ipython==8.29.0
jedi==0.19.2
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.7
langchain==0.3.7
langchain-core==0.3.18
langchain-text-splitters==0.3.2
langcodes==3.4.1
langsmith==0.1.143
language_data==1.2.0
llama-recipes==0.0.3
llama_models==0.0.50
llama_stack==0.0.52
loralib==0.1.2
lxml==5.3.0
marisa-trie==1.2.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
mccabe==0.7.0
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
multivolumefile==0.2.3
murmurhash==1.0.10
mypy-extensions==1.0.0
networkx==3.4.2
nltk==3.9.1
numpy==1.26.4
ollama==0.3.3
openai==1.39.0
opentelemetry-api==1.25.0
opentelemetry-exporter-otlp-proto-common==1.25.0
opentelemetry-exporter-otlp-proto-http==1.25.0
opentelemetry-instrumentation==0.46b0
opentelemetry-instrumentation-requests==0.46b0
opentelemetry-proto==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-semantic-conventions==0.46b0
opentelemetry-util-http==0.46b0
optimum==1.23.3
orjson==3.10.11
packaging==24.2
pandas==2.2.3
parlai==0.9.0
parso==0.8.4
pathspec==0.12.1
peewee==3.17.8
peft==0.13.2
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.3.6
pluggy==1.5.0
preshed==3.0.9
prompt_toolkit==3.0.48
propcache==0.2.0
proto-plus==1.25.0
protobuf==4.25.5
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-gfm==2.0.0
py-rouge==1.1
py7zr==0.22.0
pyaes==1.6.1
pyarrow==18.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pybcj==1.0.2
pycodestyle==2.12.1
pycryptodomex==3.21.0
pydantic==2.9.2
pydantic_core==2.23.4
pydub==0.25.1
pyflakes==3.2.0
Pygments==2.18.0
pyparsing==3.2.0
pyppmd==1.1.0
Pyrogram==2.0.106
PySocks==1.7.1
pytest==8.3.3
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-multipart==0.0.12
python-telegram-bot==21.7
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
pyzstd==0.16.2
rank-bm25==0.2.2
recommonmark==0.7.1
referencing==0.35.1
regex==2024.11.6
requests==2.32.2
requests-mock==1.12.1
requests-toolbelt==1.0.0
rich==13.5.3
rpds-py==0.21.0
rsa==4.9
ruamel.yaml==0.17.40
ruamel.yaml.clib==0.2.12
ruff==0.7.3
s3transfer==0.10.3
safehttpx==0.1.1
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
semantic-version==2.10.0
semgrep==1.96.0
sentencepiece==0.2.0
sh==2.1.0
shellingham==1.5.4
six==1.16.0
smart-open==7.0.5
smmap==5.0.1
sniffio==1.3.1
snowballstemmer==2.2.0
soupsieve==2.6
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
Sphinx==5.3.0
sphinx-autodoc-typehints==1.23.0
sphinx-rtd-theme==2.0.0
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
SQLAlchemy==2.0.27
srsly==2.4.8
stack-data==0.6.3
starlette==0.41.2
sympy==1.13.1
tabulate==0.9.0
tenacity==9.0.0
termcolor==2.5.0
texttable==1.7.0
TgCrypto==1.2.5
thinc==8.2.5
threadpoolctl==3.5.0
tiktoken==0.8.0
tokenize_rt==6.1.0
tokenizers==0.20.3
tomli==2.0.2
tomlkit==0.12.0
torch==2.5.1
tqdm==4.67.0
traitlets==5.14.3
transformers==4.46.2
typer==0.13.0
typing_extensions==4.8.0
tzdata==2024.2
tzlocal==5.2
Unidecode==1.3.8
untokenize==0.1.1
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.32.0
wasabi==1.1.3
wcmatch==8.5.2
wcwidth==0.2.13
weasel==0.4.1
websocket-client==1.8.0
websocket-server==0.6.4
websockets==12.0
wrapt==1.16.0
xxhash==3.5.0
yarl==1.17.1
zipp==3.21.0
